drivers: adxl362: rework trigger setting function
These changes fix several issues with the current implementation of the
trigger setting function.

The status bits should be cleared before mapping them to the interrupt
pin. In addition, the threshold and data ready functions can be enabled
simultaneously and the status bits should be cleared independently.

The default case of the switch statement should not jump over the
unlocking of the trigger mutex. This has been fixed with more granular
locking which has the added benefit of being more explicit about what is
being protected.

Since the trigger functions are independent, the GPIO callbacks should
be re-enabled in the error path after attempting to map the interrupts.

Signed-off-by: Brett Witherspoon <spoonb@cdspooner.com>